@@ -127,6 +127,13 @@ class Publisher(TimeStampedModel, IndestructibleModel):
         ),
     )
 
+    saas = models.BooleanField(
+        default=False,
+        help_text=_(
+            "This published is configured as a SaaS customer. They will be billed by usage instead of paid out."
+        ),
+    )
+
     # Default to False so that we can use this as an "approved" flag for publishers
     allow_paid_campaigns = models.BooleanField(_("Allow paid campaigns"), default=False)
     allow_affiliate_campaigns = models.BooleanField(
@@ -417,7 +424,7 @@ class Campaign(TimeStampedModel, IndestructibleModel):
         )
         if exclude_deprecated:
             queryset = queryset.exclude(deprecated=True)
-        return queryset
+        return queryset.distinct()
 
     def total_value(self):
         """Calculate total cost/revenue for all ads/flights in this campaign."""
@@ -1060,6 +1067,12 @@ class Advertisement(TimeStampedModel, IndestructibleModel):
             # we can't store UAs indefinitely from a user merely browsing
             user_agent = None
 
+        if div_id:
+            # Even though the publisher could have a div of any length
+            # and we want to echo back the same div to them,
+            # we only store the first 100 characters of it.
+            div_id = div_id[: Offer.DIV_MAXLENGTH]
+
         obj = model.objects.create(
             date=timezone.now(),
             publisher=publisher,
@@ -1644,6 +1657,8 @@ class AdBase(TimeStampedModel, IndestructibleModel):
 
     """A base class for data on ad views and clicks."""
 
+    DIV_MAXLENGTH = 100
+
     date = models.DateTimeField(_("Impression date"), db_index=True)
 
     publisher = models.ForeignKey(
@@ -1682,7 +1697,9 @@ class AdBase(TimeStampedModel, IndestructibleModel):
 
     # Client data
     keywords = JSONField(_("Keyword targeting for this view"), blank=True, null=True)
-    div_id = models.CharField(_("Div id"), blank=True, null=True, max_length=100)
+    div_id = models.CharField(
+        _("Div id"), blank=True, null=True, max_length=DIV_MAXLENGTH
+    )
     # This locked up the DB for a long time trying to write to our huge View table,
     # so we made it a Text field instead of a FK.
     ad_type_slug = models.CharField(_("Ad type"), blank=True, null=True, max_length=100)
